move handle_answer_modal to common.py
